requirements.txt
fastapi~=0.116.0
uvicorn[standard]~=0.35.0
pydantic~=2.11.7
streamlit~=1.46.1
requests~=2.32.4
beautifulsoup4~=4.13.4
python-dotenv~=1.1.1
google-generativeai
protobuf~=5.29.5

